requirements.txt
python==3.10.6
flask==3.0.0
win10toast==0.9
psutil==5.9.6
pyinstaller==6.2.0
windows_toasts==1.1.0
pillow==10.1.0
pyperclip~=1.8.2
pystray==0.19.5
pywin32>=311
striprtf==0.0.29
setuptools<81
zeroconf==0.148.0
customtkinter==5.2.2
